refactor method
