+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/TableOutputResolver.scala
@@ -26,7 +26,7 @@ import org.apache.spark.sql.connector.catalog.CatalogV2Implicits._
 import org.apache.spark.sql.errors.QueryCompilationErrors
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.internal.SQLConf.StoreAssignmentPolicy
-import org.apache.spark.sql.types.{ArrayType, DataType, DecimalType, IntegralType, MapType, StructType}
+import org.apache.spark.sql.types.{ArrayType, DataType, MapType, StructType}
 
 object TableOutputResolver {
   def resolveOutputColumns(
@@ -220,21 +220,6 @@ object TableOutputResolver {
     }
   }
 
-  private def containsIntegralOrDecimalType(dt: DataType): Boolean = dt match {
-    case _: IntegralType | _: DecimalType => true
-    case a: ArrayType => containsIntegralOrDecimalType(a.elementType)
-    case m: MapType =>
-      containsIntegralOrDecimalType(m.keyType) || containsIntegralOrDecimalType(m.valueType)
-    case s: StructType =>
-      s.fields.exists(sf => containsIntegralOrDecimalType(sf.dataType))
-    case _ => false
-  }
-
-  private def canCauseCastOverflow(cast: Cast): Boolean = {
-    containsIntegralOrDecimalType(cast.dataType) &&
-      !Cast.canUpCast(cast.child.dataType, cast.dataType)
-  }
-
   private def checkField(
       tableAttr: Attribute,
       queryExpr: NamedExpression,
@@ -253,11 +238,7 @@ object TableOutputResolver {
           val cast = Cast(queryExpr, tableAttr.dataType, Option(conf.sessionLocalTimeZone),
             ansiEnabled = true)
           cast.setTagValue(Cast.BY_TABLE_INSERTION, ())
-          if (canCauseCastOverflow(cast)) {
-            CheckOverflowInTableInsert(cast, tableAttr.name)
-          } else {
-            cast
-          }
+          Cast.checkOverflowInTableInsert(cast, tableAttr.name)
         case StoreAssignmentPolicy.LEGACY =>
           Cast(queryExpr, tableAttr.dataType, Option(conf.sessionLocalTimeZone),
             ansiEnabled = false)
